+++ packages/react/lib/load.ts
@@ -1,6 +1,8 @@
 import EvervaultClient, {
+  BrandOptions,
   CustomConfig as BrowserConfig,
 } from "@evervault/browser";
+import { createBrand } from "shared";
 import React, { useMemo } from "react";
 import { useId } from "react";
 
@@ -13,6 +15,13 @@ export class PromisifiedEvervaultClient extends Promise<EvervaultClient> {
   public async decrypt(token: string, data: unknown) {
     const ev = await this;
     return ev.decrypt(token, data);
+  }
+
+  get brands() {
+    return {
+      create: (name: string, options: BrandOptions) =>
+        createBrand(name, options),
+    };
   }
 }