+++ e2e-tests/inputs/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @evervault/inputs-e2e-tests
 
+## 1.0.30
+
+### Patch Changes
+
+- @evervault/inputs@2.18.32
+
+## 1.0.29
+
+### Patch Changes
+
+- @evervault/inputs@2.18.31
+
+## 1.0.28
+
+### Patch Changes
+
+- @evervault/inputs@2.18.30
+
 ## 1.0.27
 
 ### Patch Changes