+++ examples/collect-card-details/CHANGELOG.md
@@ -1,5 +1,19 @@
 # example-ui-components
 
+## 0.3.11
+
+### Patch Changes
+
+- Updated dependencies [974e08d]
+  - @evervault/js@2.13.0
+
+## 0.3.10
+
+### Patch Changes
+
+- Updated dependencies [52f1154]
+  - @evervault/js@2.12.0
+
 ## 0.3.9
 
 ### Patch Changes