+++ examples/google-pay/CHANGELOG.md
@@ -1,5 +1,19 @@
 # google-pay
 
+## 0.3.5
+
+### Patch Changes
+
+- Updated dependencies [974e08d]
+  - @evervault/js@2.13.0
+
+## 0.3.4
+
+### Patch Changes
+
+- Updated dependencies [52f1154]
+  - @evervault/js@2.12.0
+
 ## 0.3.3
 
 ### Patch Changes