+++ examples/next-3ds/CHANGELOG.md
@@ -1,5 +1,19 @@
 # example-next-3ds
 
+## 0.1.24
+
+### Patch Changes
+
+- Updated dependencies [8501902]
+  - @evervault/react@2.26.0
+
+## 0.1.23
+
+### Patch Changes
+
+- Updated dependencies [52f1154]
+  - @evervault/react@2.25.0
+
 ## 0.1.22
 
 ### Patch Changes